updated jwt secret

+++ course-02/exercises/udacity-c2-restapi/src/controllers/v0/users/routes/auth.router.ts
@@ -22,7 +22,7 @@ async function comparePasswords(plainTextPassword: string, hash: string): Promis
 }
 
 function generateJWT(user: User): string {
-    return jwt.sign(user.short(), config.jwt.secret);
+    return jwt.sign(user, config.jwt.secret);
 }
 
 export function requireAuth(req: Request, res: Response, next: NextFunction) {
@@ -82,7 +82,7 @@ router.post('/login', async (req: Request, res: Response) => {
     // Generate JWT
     const jwt = generateJWT(user);
 
-    res.status(200).send({ auth: true, token: jwt, user: user.short()});
+    res.status(200).send({ auth: true, token: jwt, user: user});
 });
 
 //register a new user
@@ -123,7 +123,7 @@ router.post('/', async (req: Request, res: Response) => {
     // Generate JWT
     const jwt = generateJWT(savedUser);
 
-    res.status(201).send({token: jwt, user: savedUser.short()});
+    res.status(201).send({token: jwt, user: savedUser});
 });
 
 router.get('/', async (req: Request, res: Response) => {
